ref: new directory for hint ui components
